refactoring about store directory structure
