+++ squre/src/store/modules/cloths/actions.ts
@@ -1,10 +1,13 @@
 import { createAsyncAction } from "typesafe-actions";
-import { ClothsDataType } from "../../Types/ContainerTypes";
+import { ClothsDataType } from "../../../Types/ContainerTypes";
 import { AxiosError } from "axios";
 
 export const GET_CLOTHS_ALL = "cloths/GET_CLOTHS_ALL";
 export const GET_CLOTHS_ALL_SUCCESS = "cloths/GET_CLOTHS_ALL_SUCCESS";
 export const GET_CLOTHS_ALL_ERROR = "cloths/GET_CLOTHS_ALL_ERROR";
+export const GET_CLOTHS_DETAIL = "cloths/GET_CLOTHS_DETAIL";
+export const GET_CLOTHS_DETAIL_SUCCESS = "cloths/GET_CLOTHS_DETAIL_SUCCESS";
+export const GET_CLOTHS_DETAIL_ERROR = "cloths/GET_CLOTHS_DETAIL_ERROR";
 
 export const getClothsAllAsync = createAsyncAction(
   GET_CLOTHS_ALL,
